+++ components/kyma-environment-broker/common/hyperscaler/azure/env.go
@@ -3,7 +3,7 @@ package azure
 import (
 	"fmt"
 
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/eu_access"
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/euaccess"
 
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/hyperscaler"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal"
@@ -21,7 +21,7 @@ func mapRegion(credentials hyperscaler.Credentials, parameters internal.Provisio
 	region := *(parameters.Parameters.Region)
 	switch parameters.PlanID {
 	case broker.AzurePlanID, broker.AzureLitePlanID:
-		if !isInList(broker.AzureRegions(eu_access.IsEURestrictedAccess(parameters.PlatformRegion)), region) {
+		if !isInList(broker.AzureRegions(euaccess.IsEURestrictedAccess(parameters.PlatformRegion)), region) {
 			return "", fmt.Errorf("supplied region \"%v\" is not a valid region for Azure", region)
 		}
