+++ components/kyma-environment-broker/internal/broker/instance_create.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/eu_access"
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/euaccess"
 
 	"k8s.io/client-go/tools/clientcmd"
 
@@ -56,7 +56,7 @@ type ProvisionEndpoint struct {
 
 	dashboardConfig dashboard.Config
 
-	euAccessWhitelist        eu_access.WhitelistSet
+	euAccessWhitelist        euaccess.WhitelistSet
 	euAccessRejectionMessage string
 
 	log logrus.FieldLogger
@@ -71,7 +71,7 @@ func NewProvision(cfg Config,
 	plansConfig PlansConfig,
 	kvod bool,
 	planDefaults PlanDefaults,
-	euAccessWhitelist eu_access.WhitelistSet,
+	euAccessWhitelist euaccess.WhitelistSet,
 	euRejectMessage string,
 	log logrus.FieldLogger,
 	dashboardConfig dashboard.Config,
@@ -277,7 +277,7 @@ func (b *ProvisionEndpoint) validateAndExtract(details domain.ProvisionDetails, 
 	// EU Access: reject requests for not whitelisted globalAccountIds
 	if isEuRestrictedAccess(ctx) {
 		logger.Infof("EU Access restricted instance creation")
-		if eu_access.IsNotWhitelisted(ersContext.GlobalAccountID, b.euAccessWhitelist) {
+		if euaccess.IsNotWhitelisted(ersContext.GlobalAccountID, b.euAccessWhitelist) {
 			logger.Infof(b.euAccessRejectionMessage)
 			err = fmt.Errorf(b.euAccessRejectionMessage)
 			return ersContext, parameters, apiresponses.NewFailureResponse(err, http.StatusBadRequest, "provisioning")
@@ -332,7 +332,7 @@ func (b *ProvisionEndpoint) validateAndExtract(details domain.ProvisionDetails, 
 
 func isEuRestrictedAccess(ctx context.Context) bool {
 	platformRegion, _ := middleware.RegionFromContext(ctx)
-	return eu_access.IsEURestrictedAccess(platformRegion)
+	return euaccess.IsEURestrictedAccess(platformRegion)
 }
 
 // Rudimentary kubeconfig validation
@@ -414,7 +414,7 @@ func (b *ProvisionEndpoint) determineLicenceType(planId string) *string {
 
 func (b *ProvisionEndpoint) validator(details *domain.ProvisionDetails, provider internal.CloudProvider, ctx context.Context) (JSONSchemaValidator, error) {
 	platformRegion, _ := middleware.RegionFromContext(ctx)
-	plans := Plans(b.plansConfig, provider, b.config.IncludeAdditionalParamsInSchema, eu_access.IsEURestrictedAccess(platformRegion))
+	plans := Plans(b.plansConfig, provider, b.config.IncludeAdditionalParamsInSchema, euaccess.IsEURestrictedAccess(platformRegion))
 	plan := plans[details.PlanID]
 	schema := string(Marshal(plan.Schemas.Instance.Create.Parameters))
