+++ components/kyma-environment-broker/internal/broker/instance_create_test.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/eu_access"
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/euaccess"
 
 	"github.com/pivotal-cf/brokerapi/v8/domain/apiresponses"
 
@@ -81,7 +81,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -154,7 +154,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -231,7 +231,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -280,7 +280,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -353,7 +353,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -428,7 +428,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -477,7 +477,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -527,7 +527,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -590,7 +590,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -653,7 +653,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -697,7 +697,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -739,7 +739,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			true,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -782,7 +782,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			true,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -823,7 +823,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -871,7 +871,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -916,7 +916,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -967,7 +967,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -1024,7 +1024,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -1081,7 +1081,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -1138,7 +1138,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{whitelistedGlobalAccountID: struct{}{}},
+			euaccess.WhitelistSet{whitelistedGlobalAccountID: struct{}{}},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -1188,7 +1188,7 @@ func TestProvision_Provision(t *testing.T) {
 			broker.PlansConfig{},
 			false,
 			planDefaults,
-			eu_access.WhitelistSet{},
+			euaccess.WhitelistSet{},
 			"request rejected, your globalAccountId is not whitelisted",
 			logrus.StandardLogger(),
 			dashboardConfig,
@@ -1292,7 +1292,7 @@ func TestRegionValidation(t *testing.T) {
 				broker.PlansConfig{},
 				false,
 				planDefaults,
-				eu_access.WhitelistSet{},
+				euaccess.WhitelistSet{},
 				"request rejected, your globalAccountId is not whitelisted",
 				logrus.StandardLogger(),
 				dashboardConfig,
