--- components/kyma-environment-broker/internal/eu_access/test/eu_access_whitelist.yaml
@@ -1,2 +1,0 @@
-whitelist:
-  - whitelisted-global-account-id
