+++ components/kyma-environment-broker/internal/euaccess/test/eu_access_whitelist.yaml
@@ -1,0 +1,3 @@
+whitelist:
+  - whitelisted-global-account-id
+  - another-whitelisted-global-account-id
