+++ components/kyma-environment-broker/internal/process/provisioning/check_cluster_configuration_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/eu_access"
+	"github.com/kyma-project/control-plane/components/kyma-environment-broker/internal/euaccess"
 
 	reconcilerApi "github.com/kyma-incubator/reconciler/pkg/keb"
 	"github.com/kyma-project/control-plane/components/kyma-environment-broker/common/gardener"
@@ -146,7 +146,7 @@ func fixOperationCreateRuntimeWithPlatformRegion(t *testing.T, planID, region, p
 			},
 		},
 	}
-	provisioningOperation.InstanceDetails.EuAccess = eu_access.IsEURestrictedAccess(platformRegion)
+	provisioningOperation.InstanceDetails.EuAccess = euaccess.IsEURestrictedAccess(platformRegion)
 	provisioningOperation.ProvisioningParameters = FixProvisioningParameters(planID, region, platformRegion)
 	provisioningOperation.RuntimeID = ""
